Create Tic-Tac-Toe
Tic-tac-toe is a game where two players X and O
 fill the hash (#) shaped box (consist of two 
vertical lines crossing two horizontal lines) 
with their alternate turns. The player who first fills 
the box with 3Xs or 3Os in a horizontal, vertical, or 
diagonal manner will win the game. 

In some cases, when none of the players succeeds in filling
the boxes horizontally, vertically, or diagonally with 3Xs 
or 3Os, then the game will be considered a draw.